BL-TBD Removed incomplete redux logic to transition this branch to another Labs participant
