+++ src/components/pages/DataVisualizations/HeatMapContainer.jsx
@@ -1,49 +1,17 @@
-import React, { useState } from 'react';
+import React from 'react';
 import WorldHeatMap from './HeatMaps/WorldHeatMap';
-import { regions } from '../../../data/filterConstants';
+import { connect } from 'react-redux';
+import { setHeatMapYear } from '../../../state/actionCreators';
 import 'antd/dist/antd.css';
 import { Select } from 'antd';
 const { Option } = Select;
-//Auto-generates a simulated API data response object
-//This schema is tentative and will need to be adjusted
-//as the backend resource and DTO schema develop
 
-const apiData = regions.map(region => {
-  return {
-    location: region.territory,
-    caseTotals: [
-      {
-        year: 2018,
-        grant: Math.floor(Math.random() * 100),
-        deny: Math.floor(Math.random() * 100),
-        dismiss: Math.floor(Math.random() * 100),
-      },
-      {
-        year: 2019,
-        grant: Math.floor(Math.random() * 100),
-        deny: Math.floor(Math.random() * 100),
-        dismiss: Math.floor(Math.random() * 100),
-      },
-      {
-        year: 2020,
-        grant: Math.floor(Math.random() * 100),
-        deny: Math.floor(Math.random() * 100),
-        dismiss: Math.floor(Math.random() * 100),
-      },
-      {
-        year: 2021,
-        grant: Math.floor(Math.random() * 100),
-        deny: Math.floor(Math.random() * 100),
-        dismiss: Math.floor(Math.random() * 100),
-      },
-    ],
-  };
+const mapStateToProps = state => ({
+  heatMapYear: state.vizReducer.heatMapYear,
+  heatMapData: state.vizReducer.heatMapData,
 });
 
-const HeatMapContainer = props => {
-  //If these components get approved, state should be moved to the Redux store
-  const [yearSelection, setYearSelection] = useState('all');
-
+const HeatMapContainer = ({ heatMapYear, heatMapData, setHeatMapYear }) => {
   //These reducer functions are passed to the component and used to process
   //incoming API data into case data arrays. Defining these functions outside
   //of the individual heatmap components allows for reusability
@@ -80,7 +48,7 @@ const HeatMapContainer = props => {
           }}
           placeholder="Filter By Office Region"
           defaultValue={'all'}
-          onChange={value => setYearSelection(value)}
+          onChange={value => setHeatMapYear(value)}
         >
           {/* Ultimately, as the data shape is better known, these should be dynamically created */}
           <Option value={'all'}>Display Year: All</Option>
@@ -98,23 +66,23 @@ const HeatMapContainer = props => {
         }}
       >
         <WorldHeatMap
-          yearSelection={yearSelection}
+          yearSelection={heatMapYear}
           reducerForm={reduceTotals}
-          apiData={apiData}
+          apiData={heatMapData}
           title={
-            yearSelection === 'all'
+            heatMapYear === 'all'
               ? 'Total Asylum Petitions 2018-2021'
-              : `Asylum Petitions In ${yearSelection}`
+              : `Asylum Petitions In ${heatMapYear}`
           }
         />
         <WorldHeatMap
-          yearSelection={yearSelection}
+          yearSelection={heatMapYear}
           reducerForm={reduceGrantRate}
-          apiData={apiData}
+          apiData={heatMapData}
           title={
-            yearSelection === 'all'
+            heatMapYear === 'all'
               ? 'Total Asylum Grant Rate % 2018-2021'
-              : `Asylum Grant Rate % For ${yearSelection}`
+              : `Asylum Grant Rate % For ${heatMapYear}`
           }
         />
       </div>
@@ -122,4 +90,4 @@ const HeatMapContainer = props => {
   );
 };
 
-export default HeatMapContainer;
+export default connect(mapStateToProps, { setHeatMapYear })(HeatMapContainer);
